Compile libpalmrejection and tests for the host
This allows us to run these tests without the device present, which
simplifies development and updates of the code base.

The approach used in libchrome to avoid the lack of "__ANDROID__"
definition is to introduce a new "__ANDROID_HOST__" definition for the
host. We follow the same convention here, which allows us to continue to
rely on libchrome.

Separately, linux uapi was not available by default for the host, and
needed to be added explicitly via 'include_dirs'.

Bug: 198472780
Test: atest libpalmrejection_test --bazel-mode --host
Change-Id: I58dcdbe3e0512b38fd7c9248c3a839b9dbe8fa14
Merged-In: I58dcdbe3e0512b38fd7c9248c3a839b9dbe8fa14
(cherry picked from commit 364e29d6fc571c4f1b5f74867379ebedabb6768f)