Time-based resampling
Currently, we are relying on the number of samples in order to do certain
actions, like checking for hold or for running the rejection model.

Before this CL, the resampling could end up producing several samples at
once. This would cause the actions above to potentially be skipped,
because we performed the check of the number of samples before the
resampling code ran. An exact match for certain number of samples was
performed, so sometimes, palm rejection model would not get executed.

In this CL, we change the approach to rely on the time instead. We check
whether a specific time threshold is crossed with each sample received
from the touchscreen. If it's crossed, then the action is taken.

Additional changes here:
1) All samples are now stored in unmodified manner.
2) It's possible to exceed max_sample_count when resampling is enabled
3) samples_seen() is replaced with Duration()
4) When a resampled value is needed, we get getSampleAt(time). The
   resampled value is never stored.
5) Old variables such as last_sample and old function Resample were
   deleted
6) We are now storing the first sample's timestamp in first_sample_time.
   It's needed to calculate the duration of the stroke, since at some
   point, we will drop the first sample

After this change, any references to samples_seen and samples().size()
should consider the resampled and non-resampled cases.
Therefore, the "Filter" function should not directly check the number of
samples anywhere.

Bug: b/240168494
Test: atest libpalmrejection_test inputflinger_tests
Change-Id: I983aa93ce7ae352aca748d4bbab8fd2d1af28df5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3857630
Commit-Queue: Xinglong Luan <alanlxl@chromium.org>
Reviewed-by: Rob Schonberger <robsc@chromium.org>
Auto-Submit: Siarhei Vishniakou <svv@google.com>
Reviewed-by: Xinglong Luan <alanlxl@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1049397}